scripts: add Zephyr inline code generation
Codegen is a file generation tool. It use pieces of Python code in the
source file as generators to generate whatever text is needed.

If a source file is added by  zephyr_sources_codegen(..) or
zephyr_sources_codegen_ifdef(..) codegen will be executed and
the generated file stored to the build directory. Compilation
will run on the generated file only.

Signed-off-by: Bobby Noelte <b0661n0e17e@gmail.com>