drivers/flash: qspi_stm32: Move to dma support
Move from IT based to DMA based QSPI data transfer.
Due to use of the HAL by this driver, the DMA API usage required
some adaptation as DMA HAL is called directly within QSPI HAL
and can't be avoided.
As a consequence both HAL and Zephyr dma api should be initialized:
-HAL DMA API because it requires knowledge of the transfers to be done
-Zephyr DMA API because we need to properly route the IRQ callbacks

Since Zephyr DMA API does not support such case, a particular unused
field of the API is set to a particular value to inform the STM32
DMA driver of this override by the HAL.

Signed-off-by: Erwan Gouriou <erwan.gouriou@linaro.org>