drivers/dma: stm32: Deal with half transer IRQ
In case half transfer IRQ has been configured (as done by HAL API
in case of QSPI transfer), treat this flag before transfer complete,
as it might occur that both are set at same time.

Additionally clear the flags after the callback is treated, so it
could be treated in HAL IRQ handler.

Signed-off-by: Erwan Gouriou <erwan.gouriou@linaro.org>